# swagger.json
{
  "openapi": "3.0.4",
  "paths": {
    "/api/bikers": {
      "post": {
        "operationId": "BikersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBikerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Biker"
                }
              }
            }
          }
        },
        "tags": [
          "Bikers"
        ]
      }
    },
    "/api/bikers/{id}/confirm": {
      "get": {
        "operationId": "BikersController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Bikers"
        ]
      }
    },
    "/api/bikers/{id}": {
      "patch": {
        "operationId": "BikersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBikerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Biker"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Bikers"
        ]
      }
    },
    "/api/bikers/{id}/credit-cards": {
      "post": {
        "operationId": "BikersController_changeCreditCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCreditCardDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Bikers"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/rentals": {
      "post": {
        "operationId": "RentalsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          }
        },
        "tags": [
          "Rentals"
        ]
      }
    },
    "/api/rentals/return": {
      "post": {
        "operationId": "RentalsController_postReturn",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          }
        },
        "tags": [
          "Rentals"
        ]
      }
    }
  },
  "info": {
    "title": "Bike Sharing System API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateCreditCardDto": {
        "type": "object",
        "properties": {
          "creditCardNumber": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string"
          }
        },
        "required": [
          "creditCardNumber",
          "expirationDate"
        ]
      },
      "CreatePassportDto": {
        "type": "object",
        "properties": {
          "expirationDate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "expirationDate"
        ]
      },
      "CreateBikerDto": {
        "type": "object",
        "properties": {
          "cpf": {
            "type": "string"
          },
          "birthDate": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          },
          "confirmationPassword": {
            "type": "string"
          },
          "creditCard": {
            "$ref": "#/components/schemas/CreateCreditCardDto"
          },
          "passport": {
            "$ref": "#/components/schemas/CreatePassportDto"
          }
        },
        "required": [
          "birthDate",
          "email",
          "password",
          "confirmationPassword",
          "creditCard"
        ]
      },
      "CreditCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "creditCardNumber": {
            "type": "string"
          },
          "holderName": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "creditCardNumber",
          "holderName",
          "expirationDate"
        ]
      },
      "Passport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "passportNumber": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "expirationDate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "passportNumber",
          "countryCode",
          "expirationDate"
        ]
      },
      "Station": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "stationSerial": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "docks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dock"
            }
          }
        },
        "required": [
          "id",
          "stationSerial",
          "name",
          "location",
          "docks"
        ]
      },
      "Dock": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "dockSerial": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "manufactureDate": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPERATIONAL",
              "AVAILABLE",
              "OCCUPIED",
              "MAINTENANCE_REQUESTED",
              "UNDER_MAINTENANCE",
              "DECOMMISSIONED"
            ]
          },
          "bike": {
            "$ref": "#/components/schemas/Bike"
          },
          "station": {
            "$ref": "#/components/schemas/Station"
          },
          "rentals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            }
          },
          "returns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            }
          }
        },
        "required": [
          "id",
          "dockSerial",
          "model",
          "manufactureDate",
          "status"
        ]
      },
      "Bike": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bikeSerial": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "manufactureYear": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "AVAILABLE",
              "RENTED",
              "MAINTENANCE_REQUESTED",
              "UNDER_MAINTENANCE",
              "RETIRED"
            ]
          },
          "dock": {
            "$ref": "#/components/schemas/Dock"
          },
          "rentals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            }
          }
        },
        "required": [
          "id",
          "bikeSerial",
          "brand",
          "model",
          "manufactureYear",
          "status"
        ]
      },
      "Rental": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "finishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "biker": {
            "$ref": "#/components/schemas/Biker"
          },
          "bike": {
            "$ref": "#/components/schemas/Bike"
          },
          "rentedFromDock": {
            "$ref": "#/components/schemas/Dock"
          },
          "retunedToDock": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Dock"
              }
            ]
          },
          "initialCharge": {
            "$ref": "#/components/schemas/Charge"
          },
          "extraCharge": {
            "$ref": "#/components/schemas/Charge"
          }
        },
        "required": [
          "id",
          "startedAt",
          "finishedAt",
          "biker",
          "bike",
          "rentedFromDock",
          "retunedToDock",
          "initialCharge"
        ]
      },
      "Biker": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "ACTIVE"
            ],
            "type": "string"
          },
          "creditCard": {
            "$ref": "#/components/schemas/CreditCard"
          },
          "passport": {
            "$ref": "#/components/schemas/Passport"
          },
          "charges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Charge"
            }
          },
          "rentals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            }
          }
        },
        "required": [
          "id",
          "name",
          "birthDate",
          "email",
          "password",
          "status",
          "creditCard",
          "charges",
          "rentals"
        ]
      },
      "Charge": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "requestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "default": 10
          },
          "biker": {
            "$ref": "#/components/schemas/Biker"
          },
          "rental": {
            "$ref": "#/components/schemas/Rental"
          },
          "extra": {
            "$ref": "#/components/schemas/Rental"
          }
        },
        "required": [
          "id",
          "requestedAt",
          "completedAt",
          "amount",
          "biker",
          "rental",
          "extra"
        ]
      },
      "UpdateBikerDto": {
        "type": "object",
        "properties": {
          "confirmationPassword": {
            "type": "string"
          },
          "birthDate": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          },
          "passport": {
            "$ref": "#/components/schemas/CreatePassportDto"
          }
        }
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}